#17 - Changes in tests and buy/sell endpoints. Add delete order and integration tests for delete order.

+++ ExchangeSimulator/ExchangeSimulator.Application/Requests/PlayerRequests/GetMyPlayer/GetMyPlayerDto.cs
@@ -31,7 +31,7 @@ public class GetMyPlayerDto
     /// <summary>
     /// Total volume
     /// </summary>
-    public int TurnOver { get; set; }
+    public decimal TurnOver { get; set; }
 
     /// <summary>
     /// List of players coins
